Moved sqllab tests from core_tests to sqllab_tests
